# docs/swagger/swagger.yaml
basePath: /walm/api/v1
definitions:
  cluster.Cluster:
    properties:
      apps:
        items:
          type: '&{helm ReleaseRequest}'
        type: array
      configvalues:
        type: object
    type: object
  ex.ApiResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  helm.ReleaseInfo:
    properties:
      chartappversion:
        type: string
      chartname:
        type: string
      chartversion:
        type: string
      configvalues:
        type: object
      dependencies:
        type: object
      name:
        type: string
      namespace:
        type: string
      statuscode:
        type: integer
      version:
        type: integer
    type: object
  helm.ReleaseRequest:
    properties:
      chartname:
        type: string
      chartversion:
        type: string
      configvalues:
        type: object
      dependencies:
        type: object
      name:
        type: string
      namespace:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: bing.han@transwarp.io
    name: bing.han
    url: http://transwarp.io
  description: cluster lifecycle manager
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Walm
  version: 1.0.0
paths:
  /cluster/namespace/{namespace}/name/{name}:
    delete:
      consumes:
      - application/json
      description: Delete an Cluster
      parameters:
      - description: identifier of the namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: the name of cluster
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: cluster not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - Cluster
    get:
      consumes:
      - application/json
      description: Get states of an Cluster
      parameters:
      - description: identifier of the namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: the name of cluster
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/helm.ReleaseInfo'
            type: array
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: cluster not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - Cluster
    post:
      consumes:
      - application/json
      description: Deploy an Cluster
      parameters:
      - description: identifier of the namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: the name of cluster
        in: path
        name: name
        required: true
        type: string
      - description: Apps of Cluster
        in: body
        name: apps
        required: true
        schema:
          $ref: '#/definitions/cluster.Cluster'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: namespace not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - Cluster
  /cluster/namespace/{namespace}/name/{name}/instance:
    post:
      consumes:
      - application/json
      description: Deploy an Instance into Cluster
      parameters:
      - description: identifier of the namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: the name of cluster
        in: path
        name: name
        required: true
        type: string
      - description: Apps of Cluster
        in: body
        name: apps
        required: true
        schema:
          $ref: '#/definitions/helm.ReleaseRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: namespace not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - Cluster
  /instance/namespace/{namespace}/list:
    get:
      consumes:
      - application/json
      description: find the aim instance status
      parameters:
      - description: identifier of the instance
        in: path
        name: namespace
        required: true
        type: string
      - description: identifier of the appname
        in: path
        name: appname
        required: true
        type: string
      - description: max num to display
        in: query
        name: max
        type: integer
      - description: the offset of result
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/helm.ReleaseInfo'
            type: array
        "404":
          description: Invalid status not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - instance
  /instance/namespace/{namespace}/name/{appname}:
    delete:
      consumes:
      - application/json
      description: Delete an Appliation
      parameters:
      - description: identifier of the instance
        in: path
        name: namespace
        required: true
        type: string
      - description: identifier of the instance
        in: path
        name: appname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - instance
    post:
      consumes:
      - application/json
      description: deploy an instance with the givin data
      parameters:
      - description: identifier of the instance
        in: path
        name: namespace
        required: true
        type: string
      - description: identifier of the instance
        in: path
        name: instance
        required: true
        type: string
      - description: ReleaseRequest of instance
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/helm.ReleaseRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "405":
          description: Invalid input
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - instance
    put:
      consumes:
      - application/json
      description: Update an Appliation
      parameters:
      - description: namespace of the instance
        in: path
        name: namespace
        required: true
        type: string
      - description: identifier of the instance
        in: path
        name: appname
        required: true
        type: string
      - description: Update instance
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/helm.ReleaseRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "405":
          description: Invalid input
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - instance
  /instance/namespace/{namespace}/name/{appname}/info:
    get:
      consumes:
      - application/json
      description: Get an Appliation by name with status
      parameters:
      - description: identifier of the instance
        in: path
        name: namespace
        required: true
        type: string
      - description: identifier of the application
        in: path
        name: appname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/helm.ReleaseInfo'
            type: array
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "405":
          description: Invalid input
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - instance
  /instance/namespace/{namespace}/name/{appname}/version/{version}/rollback:
    get:
      consumes:
      - application/json
      description: Rollback Instance to aim version
      parameters:
      - description: identifier of the instance
        in: path
        name: namespace
        required: true
        type: string
      - description: identifier of the instance
        in: path
        name: appname
        required: true
        type: string
      - description: identifier of the version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "400":
          description: Invalid Name supplied!
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "405":
          description: Invalid input
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/ex.ApiResponse'
            type: object
      tags:
      - instance
swagger: "2.0"