# router/api/v1/cluster/cluster.go
package cluster

import (
	helm "walm/pkg/helm"
)

type Cluster struct {
	ConfigValues map[string]interface{} `json:"configvalues" description:"extra values added to the chart"`
	Apps         []helm.ReleaseRequest  `json:"apps" type:"array" items:"helm.ReleaseRequest"  description:"list of application of the cluster"`
}